1.1 release 2.
Release notes:

Hello, hello!

Feeling was submitted to the app store this weekend, but I pulled the
release on Monday and decided to include some updates on the initial
release to the App Store.

I'm pretty excited about this one because I think it brings the app to a
whole new level of usefulness.

New features:

- Events now have **image** and **note** fields. Hooray!
- The table view for viewing old events has been updated and is a lot more useful.
- Tapping events shows a special page for viewing old events, as well as their notes and photos.

Really excited about this update and after a couple small bug fixes
here and there, we'll be on our way to initial release.

<3